[FastRoute] Move public headers to correct subdir
